modules/game-engine/requirements.txt
# imageio==2.31.1
# keyboard==0.13.5
matplotlib==3.7.1
numpy==1.24.4
# opencv-python-headless==4.7.0.72
paho-mqtt==1.6.1
transitions==0.9.2
